- Fixed some compile-time warnings. Could fix crashes on Windows (especially ones about not returning a value). - Fixed linking problems.

+++ trunk/src/ageometry-svn/CSceneNodeAnimatorCameraFOV.h
@@ -16,7 +16,7 @@ namespace irr
     public:
 
       //! constructor
-      CSceneNodeAnimatorCameraFOV(f32 fovSpeed=0.000001);
+      CSceneNodeAnimatorCameraFOV(f32 fovSpeed=0.000001f);
 
       //! destructor
       virtual ~CSceneNodeAnimatorCameraFOV();
@@ -48,7 +48,7 @@ namespace irr
       f32 _maxFOV;
       f32 _minFOV;
       
-      f32 KeyControl[2];
+      u32 KeyControl[2];
       u32 lastUpdate;
     };
